+++ migrations.go
@@ -88,6 +88,7 @@ var (
 		addDefaultACSRGateOptionsToChampionships,
 		addSplitTypeToRaceWeekends,
 		fixCarDuplicationInRaceSetups,
+		addRealPenaltyAppUDPPort,
 	}
 )
 
@@ -1180,4 +1181,18 @@ func fixCarDuplicationInRaceSetups(s Store) error {
 	}
 
 	return nil
+}
+
+func addRealPenaltyAppUDPPort(s Store) error {
+	logrus.Infof("Running migration: Add Real Penalty App UDP Port")
+
+	rpOpts, err := s.LoadRealPenaltyOptions()
+
+	if err != nil {
+		return err
+	}
+
+	rpOpts.RealPenaltyAppConfig.General.AppUDPPort = rpOpts.RealPenaltyAppConfig.General.AppTCPPort
+
+	return s.UpsertRealPenaltyOptions(rpOpts)
 }