+++ plugin_realpenalty_config.go
@@ -46,6 +46,7 @@ func DefaultRealPenaltyAppConfig() RealPenaltyAppConfig {
 			UDPPort:           0,
 			UDPResponse:       "",
 			AppTCPPort:        53000,
+			AppUDPPort:        53000,
 		},
 		PluginsRelay: RealPenaltyPluginsRelay{
 			OtherUDPPlugin: "",
@@ -109,6 +110,7 @@ type RealPenaltyConfigGeneral struct {
 	UDPPort     int    `ini:"UDP_PORT" show:"-" help:"Listening UDP port - Set the same port (without IP) of cfg of the server, UDP_PLUGIN_ADDRESS "`
 	UDPResponse string `ini:"UDP_RESPONSE" show:"-" help:"Destination IP and UDP port for response - Set the same port in the cfg of the AC server, UDP_PLUGIN_LOCAL_PORT"`
 	AppTCPPort  int    `ini:"APP_TCP_PORT" show:"open" help:"Listening TCP port from AC app (to open in firewall/router). Must be one of 53000, 53001, 53002 to 53020, or the port of cfg AC server, HTTP_PORT + 27. The app will try all these ports on the ac server's ip address (until the right connection is found)"`
+	AppUDPPort  int    `ini:"APP_UDP_PORT" show:"open" help:"Listening UDP port from AC app (to open in firewall/router). Any free UDP port is OK"`
 
 	AppFile      string `ini:"APP_FILE" show:"-" help:"Path and file names of the app from the plugin package"`
 	ImagesFile   string `ini:"IMAGES_FILE" show:"-" help:"Path and file names of the images from the plugin package"`